pyproject.toml
[project]
name = "console-based-agentic-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dataclasses>=0.8",
    "openai-agents>=0.2.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
